requirements.txt

Authlib==1.6.0
ffmpeg-python==0.2.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.24.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
openai-whisper==20250625
pandas==2.3.1
pillow==11.3.0
pydub==0.25.1
regex==2024.11.6
requests-oauthlib==2.0.0
rsa==4.9.1
streamlit==1.47.0
tiktoken==0.9.0
yt-dlp==2025.6.30
torch==2.7.1
pyaudio
audioop-lts
supabase==2.17.0
